ml-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.2.0
opencv-python-headless==4.9.0.80
torch==2.2.0
torchvision==0.17.0